group players by their nationalities and the past careers to better show cohesion
The outermost circle represents nationality, and the color is based on the national flag color. In terms of ages, we visualize them through the line. This is also an age baseline circle.A player is relatively old if his age line crosses the baseline.We may also see the overall team age through these lines.